requirements.txt
aenum==3.1.11
aiohttp==3.8.3
aiohttp-cors==0.7.0
aiosignal==1.3.1
anyio==3.6.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
devtools==0.10.0
distlib==0.3.4
dnspython==2.3.0
envparse==0.2.0
executing==1.2.0
filelock==3.7.1
fonttools==4.38.0
frozenlist==1.3.3
gobject==0.1.0
graphql-core==2.3.2
graphql-server-core==1.2.0
htmlmin==0.1.12
idna==3.3
ImageHash==4.3.1
Jinja2==3.1.2
joblib==1.2.0
kaggle==1.5.12
kiwisolver==1.4.4
lark==1.1.5
MarkupSafe==2.1.2
matplotlib==3.6.3
motor==3.1.1
multidict==6.0.3
multimethod==1.9.1
networkx==3.1
numpy==1.23.5
packaging==23.0
pandas==1.5.3
patsy==0.5.3
phik==0.12.3
Pillow==9.4.0
platformdirs==2.5.2
promise==2.3
pycairo==1.21.0
pycodestyle==2.10.0
pycryptodomex==3.16.0
pydantic==1.10.7
pygccxml==2.2.1
Pygments==2.14.0
PyGObject==3.42.2
pygtrie==2.5.0
pymongo==4.3.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-ndn==0.3.post2
python-slugify==6.1.2
PyTrie==0.4.0
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.1
Rx==1.6.3
scapy==2.5.0
scipy==1.9.3
seaborn==0.12.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
statsmodels==0.13.5
tangled-up-in-unicode==0.2.0
text-unidecode==1.3
tomli==2.0.1
tqdm==4.64.0
typeguard==2.13.3
typing_extensions==4.5.0
urllib3==1.26.11
visions==0.7.5
yarl==1.8.2
ydata-profiling==4.1.2
